Type:chore(cppcheck)
Description:

RootCause:Null

Solution:
Reviewed-on: http://gerrit.uniontech.com/c/chengdu/dde-file-manager/+/7599
Tested-by: <mailman@uniontech.com>

+++ dde-file-manager-plugins/pluginPreview/dde-music-preview-plugin/toolbarframe.cpp
@@ -139,7 +139,7 @@ void ToolBarFrame::onPlayControlButtonClicked()
 
 void ToolBarFrame::updateProgress()
 {
-    m_progressSlider->setValue(m_player->position());
+    m_progressSlider->setValue(static_cast<int>(m_player->position()));
 }
 
 void ToolBarFrame::seekPosition(const int &pos)
